Extract pairing diagnostics
